+++ bindings/java/iota-wallet-java/CHANGELOG.md
@@ -19,6 +19,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.0.0-rc.3 - XXXX-XX-XX
+
+### Added
+
+- Enable native library loading via Java path;
+
+### Changed
+
+- Rename `ClientCommand` to `WalletCommand`;
+- Move `WalletCommand` class from `NativeApi`;
+- Rename `ClientResponse` to `WalletResponse`;
+
 ## 1.0.0-rc.2 - 2022-11-15
 
 ### Changed